+++ client/src/index.css
@@ -55,7 +55,8 @@ h1 {
   position: absolute;
   top: 50%;
   left: 50%;
-  transform: translate(-50%, -50%);
+  width: 75%;
+  transform: translate(-47.5%, -50%);
 }
 
 .hidden {
